token_flow_project/requirements.txt
crewai>=0.28.0
crewai-tools>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
pyyaml>=6.0
matplotlib>=3.7.0
requests>=2.28.0
